Add a sender to the AtspiEvent struct.
Normally, this is set to the application / root accessible from the
application corresponding to the event source, but screen-reader-generated
events are actually generated by the screen reader. The sender is set based on
the application that sent the signal over dbus, so, in the latter case, this
will be set to the screen reader.

Per discussion at https://gitlab.gnome.org/GNOME/orca/issues/36