Defer application registration via an idle callback
A python application, for instance, may initialize gtk but not enter the
main loop. It would, thus, register with the AT-SPI registry but not be
responsive to queries. Putting registration into an idle callback ensures
that, if a program never enters the main loop, then it won't register.

Fixes https://gitlab.gnome.org/GNOME/at-spi2-core/issues/16