atkplug: Add atk_plug_set_child
As discussed on

https://mail.gnome.org/archives/gnome-accessibility-devel/2019-August/msg00001.html

In some cases, one can not use the AtkPlug type directly as accessible
object for the toplevel widget of the application. For instance in the gtk
case, GtkPlugAccessible can not inherit both from GtkWindowAccessible and
from AtkPlug. In such a case, one can create, in addition to the standard
accessible object for the toplevel widget, an AtkPlug object, and make the
former the child of the latter by calling atk_plug_set_child().