Add ATK_TEXT_ATTR_TEXT_POSITION to the list of possible AtkText attributes
IAccessible2 has a "text-position" attribute with values of "baseline",
"super", and "sub". This relative position is arguably preferable to
having applications specify the pixels via ATK_TEXT_ATTR_RISE, which
applications are not using anyway. Adding text position as an officially
supported attribute increases the likelihood that assistive technologies
will get this information in a predictable fashion.