atspi_accessible_set_cache_mask: relax assert
Atspi_accessible_set_cache_mask checks that it is being passed an
application's root accessible, and this sometimes fails and generates a
warning similar to the following, even when orca passes what it sees as the
application's root accessible:

(orca:18809): dbind-CRITICAL **: 09:27:25.951: atspi_accessible_set_cache_mask: assertion 'accessible == accessible->parent.app->root' failed

I suspect that this is caused by atk-bridge being initialized and caching
the root accessible and atk_get_root() is subsequently redefined, in the case
of an application that loads more than one atk implementor, such as Firefox.
This is indicative of a problem that should be fixed elsewhere--possibly
gtk+ and/or firefox shouldn't override atk_get_root if it already has an
implementation--but warning and failing here isn't helpful.