bus: Set X root property when Xwayland starts on-demand
On wayland, the X11 server may not always be running so that
at-spi-bus-launcher cannot set it when it is starting.

This is usually not a big problem, as applications can retrieve the a11y
bus address by querying the session bus. Unfortunately, there may be
exceptions such as applications that are run as a different user.

We can easily solve this problem by setting the X11 root property when
the Xwayland server is started. To do so, install a simple script to
/etc/xdg/Xwayland-session.d/ that queries the session bus and sets the
root property.

This script is only needed when Xwayland is started on-demand, which can
only happen on systemd enabled sessions. However, it is always safe to
install, and there will be no severe side effects should the script be
running needlessly or failing.

Closes: #34