Remove unused function setup_bus_child_daemon()
The switch to G_SPAWN_LEAVE_DESCRIPTORS_OPEN makes
g_spawn_async_with_pipes_and_fds() use the code path for
posix_spawn(), which doesn't use a child setup func.

I think setting the bus to exit when at-spi-bus-launcher exits is a
leftover from when the latter was rewritten in 2011, and we didn't
have reliable ways of ending session daemons.