First cut at building with Freedesktop CI templates - images for openSUSE and Fedora
The Fedora build does not run the tests yet.  Preferably I'd like it
to use systemd socket activation for dbus-broker, but this is proving tricky.

The Fedora build has the meson flags from the Fedora specfile for
at-spi2-core, but builds without dbus-daemon with the intention of
really testing dbus-broker eventually.