Fix GetActions
The previous code created a (sss) structure and tried to insert four strings.
Ideally, the action's unlocalized name would be sent in the structure. But
this would require changing the xml specification, which would break gtk 4's
implementation. For now, do what gtk 4 does and send the localized name,
description, and keybinding for each action.

Helps #20