xml: Remove directions from signal arguments
Per the dbus spec, specifying a direction for signal arguments is not
required, since there is only one possible direction. If a direction is
specified at all, then it is expected to be "out". Previously, the direction
was set to "in" in most cases, causing d-feet to complain. Let's just remove
the direction annotations.