Hack to make the shell pipeline work for grepping coverage.xml
The first call to grep actually finds what it is supposed to
find (lines with 'line-rate' and numbers), but it exits with a nonzero
exit code.  The rest of the pipeline is fine.

I have no idea why that first grep is doing that.  So, hack around
it.  The end result is the correct one, anyway.