mock-gnome-session.py: New mock for the org.gnome.SessionManager DBus interface
This uses python-dbusmock: https://github.com/martinpitt/python-dbusmock

Which is an *awesome* tool for creating mocks of DBus interfaces.

This mock is for the basic functionality of the
org.gnome.SessionManager and org.gnome.SessionManager.ClientPrivate
interfaces.  With it, applications can register as clients of the
session manager and get notified when the SM gets a Logout command.

My intention is to polish this a bit and finally submit it as a
built-in template to python-dbusmock; it already contains mock
templates for upower, gnome-screensaver, and other goodies.