run-tests.sh: Launch a mock org.gnome.SessionManager to control the lifetime of the daemons
Both at-spi-bus-launcher and at-spi2-registryd will keep running
indefinitely unless the session manager tells them that the session is
ending.

During the CI tests, the daemons are killed when the tests finish.
They don't get a chance to exit cleanly.

GCC's code coverage machinery from libgcc depends on a process exiting
cleanly for it to be able to write out the .gcda files.  These files
were never written for the a11y daemons because they didn't exit
cleanly.

This commit changes the test runner to launch a mock implementation of
the org.gnome.SessionManager interface.  It does nothing but accept
clients that register themselves, and to signal clients for session
termination.

The daemons now pick up that signal from the mock, as if it were from
a real session, and exit cleanly.