test-application.c: Install a signal handler to exit cleanly
The app-test program gets run by atk_test_util.c, and gets sent a
SIGTERM when the parent process wants it to terminate.  This is fine,
but if app-test terminates forcefully with the default handler for
SIGTERM, then libgcc won't get a chance to write out its code coverage
information.

So, we install a signal handler for SIGTERM, and the handler exits the
main loop.