XML: Remove the toplevel node name
https://dbus.freedesktop.org/doc/dbus-specification.html#introspection-format

Per the DBus spec, "Only the root <node> element can omit the node
name, as it's known to be the object that was introspected."

When we have a DBus XML description:

  <node>
    <interface name="Foo">
    </interface>
  </node>

The toplevel <node> element does not need a `name` attribute.

While we are at it, reindent all the XML files properly.