atk_test_selection_get_selected_child: let the compiler know that the counter won't go past the limit
The selected_count of children must be equal to the number of objects
in test-selection.xml that are marked as selected.  We know that
number, but the compiler doesn't - so, put in an assertion to let it
know that the counter won't go past the length of the number of
available children.

Caught by the static-scan job.