Control the registry fixture's lifetime via a mock gnome-session
As for the code coverage job, we need registryd's tests to be able to
control the lifetime of at-spi-bus-launcher and at-spi2-registryd.

So, the session_manager() fixture assumes that it is running alonside
a gnome-session dbus mock.  The teardown code tells the session to
Logout - this will make the mock notify session clients that they have
to terminate.