Replace all the teardown functions with a single one
For the at-spi2-atk tests, we'll move to having a test fixture that
consists of a launched test app with its claimed dbus name.  The
teardown step will terminate the test app.

In this commit, terminate_app() is renamed to fixture_teardown().  It
does not use its arguments yet, because it only uses global variables
from atk_test_util.c, but it will soon.