SpiDEController: don't push the controller itself into the message_queue
These two pointers were pushed at the same time into controller->message_queue:

  message to be processed later
  user_data / controller

And later they were popped to handle the message.  But since we now
weave the controller via the user_data everywhere, there's no need to
keep an extra copy in the message_queue.

This also leaves homogeneous elements, of the same type, in the
message_queue, for peace of mind.