atspi_register_device_event_listener(): turn into a no-op.
This function is never called from pyatspi2 nor from orca.  It was
meant to be the way to register for mouse events, but that is done
with the Registry.RegisterEvent("mouse", ...) method instead.

I've turned the function into a no-op to avoid ABI breakage.  We can
remove it when we have an ABI break.