spi_controller_notify_mouselisteners(): Remove
This function immediately returns FALSE if there are no
controller->mouse_listeners.

The only place that ever added to the controller->mouse_listeners list
was spi_controller_register_device_listener(), which we removed in the
last commit.

So, remove the function altogether and make the two callers do as if
it just returned FALSE.