Reformat all the *.[ch] files with clang-format
I ran this on each directory with C files:

  clang-format -i *.[ch]

"-i" is the in-place option.

I also adjusted the order of #includes for some files which failed to
build after that:

Clang-format reorders blocks of #include directives alphabetically,
but they can be grouped and separated by blank lines.  If there is a
blank line between blocks, like

  #include "zork.h"

  #include "bar.h"
  #include "foo.h"

then it will not put zork.h after the other two.  The last two header
files will be sorted alphabetically.

We can adjust the formatting of chunks of code by hand with comments
like these:

  /* clang-format off */

     this code {
       is, formatted, by, hand;
     }

  /* clang-format on */

See https://clang.llvm.org/docs/ClangFormat.html for the general
manual and https://clang.llvm.org/docs/ClangFormatStyleOptions.html
for the style options and the comments described above.