tests: use "gdbus wait" after launching the mocks instead of sleeping
On a busy system (the CI runners) or one with a cold cache (my machine
when running the tests after just having started the CI container
image), the "sleep 1" we had in the test runner scripts is not a
reliable way for the python/dbusmock process to have acquired the name
it is supposed to acquire.

So, use "gdbus wait" instead as a reliable way of doing that.

This should fix the spurious test failures we've been having in CI.