x11: Register key grabs using the currently-focused window
Previously, the root window was being passed to XIGrabKeycode. This had the
annoying side effect of briefly taking focus from the window when one of our
key grabs was used, leading to a window:deactive event that would confuse
orca.

The challenge with the new approach is that we need to reset any active grabs
when the focus changes. Now the X event watcher will listen for FocusIn
events and refresh the key grabs when it sees one. As a backup, we also
listen for window:activate and window:deactivate events and also refresh the
key grabs when we see one.

Fixes #101