Use internal names when possible for localized role names
This matches the behavior of atspi_accessible_get_role_name. We already hav
translations from atk, so we can use those.

Also, move the table for role name translations into atspi, since it seems
that atk is increasingly becoming a legacy tool.

Add a note to the xml indicating that it is optional to implement
GetRoleName and GetLocalizedRoleName.

Fixes #110