Remove code related to sending properties with events
This was intended to provide a way for  ATs to proactively request the
information that they were going to  need when a provider sent an event, in
order to reduce the number of round-trip dbus calls, but it has likely never
been used, and it complicates the interface. The a{sv} will be removed from
the events in the spec in a future commit.