Remove spurious hand-made deprecation notice in the documentation
This probably predates gtk-doc support for making things as deprecated,
and have been copied over inadvertently in some cases.  Nowadays the
documentation uses the proper markers anyway.

* atspi_text_get_attributes() is indeed deprecated, and marked as such.
* atspi_text_get_text_attributes() is not actually deprecated.
* atspi_role_get_name() is (probably?) not actually deprecated.

The case of atspi_role_get_name() is a bit less clear as it has had
this marker for a long time, but its feature are used by Orca at least
and there are no current replacement.