AtspiDeviceLegacy: consume mapped modifier keys when pressed alone
If a key has been mapped using atspi_device_map_modifier, then assume that it
is being used as the screen reader modifier key. As such, inform
at-spi2-registryd that it should be consumed, unless the key is released and
pressed again quickly (within 0.5 seconds). This makes the legacy back end's
behavior more closely resemble the existing behavior of orca in terms of
flagging caps lock presses to be consumed when it is set as the orca modifier.