Add AtspiLive enum, and use for Announcement events
Announcement events were missing a politeness level. Document that this should
be specified in the detail1 field. Also, add a "notification" signal to
AtkObject, deprecate the announcement signal that lacks the politeness
indicator, and map the new notification event to the AT-SPI Announcement
signal. Ideally, the new field would be simply added to the existing
announcement signal in atk, but this would make atk 2.50+ incompatible with
2.46 and 2.48, which would likely have caused issues if a binary was built
with a mismatched atk version.

Fixes #135