collection: Avoid locking up if an object has a very large child count
LibreOffice's calculator's table object currently reports 2^32 children,
causing it to lock up when the collection API queries the table. So set a
limit on the number of children that will be queried. Children beyond this
point will be ignored. Currently, this is set at 65536.

An alternate approach would be to avoid querying accessibles that have
ATSPI_STATE_MANAGES_DESCENDANTS set, but this would also affect some
well-behaved accessibles.

Fixes #138.