AtspiDeviceX11: rework the way we filter duplicate XI_KeyPress events
When a key is grabbed, we receive two events for it but only want to make
one call to atspi_device_notify_key. The old code for this was complicated
and brittle and didn't always work correctly (ie,
https://gitlab.gnome.org/GNOME/orca/-/issues/391). Instead of trying to
store an expected deviceid, compare each event's deviceid to the event's
sourceid, and filter the event when they do not match.