Have atspi_device_add_key_grab return 0 if XIGrabKeycode fails
Also refactor internally a bit: make it possible for a class handler to
indicate a failure, and always pass grabs on to the class handler,
regardless of whether they are considered duplicates. The latter simplifies
the code a bit and should also help in the future if we have cases where we
wish to chain devices together. The X11 code already has checks for
duplicate grabs, so this should be safe.

* Helps #147