Add atspi_device_generate_mouse_event, and implement for mutter
The original atspi_generate_mouse_event took absolute screen coordinates,
which applications cannot provide under Wayland. The new function takes an
Accessible along with x/y coordinates relative to that accessible, and it
can support platform-specific back ends for the implementation. For mutter,
we start a screen-cast session and use its libei interface to control the
mouse. We need the gnome-shell Introspect interface to map AT-SPI windows
to window IDs that can be sent to mutter. This interface is locked down, so
unsafe mode currently needs to be set for any of this to work. Ie, press
alt-f2, then type lg to open the looking glass, then enter
global.context.unsafe_mode = true

Closes #158