Back out code to generate mouse events on mutter
The code in atspi-device-mutter.c sent mouse events through the back door
using mutter's screen-cast interface, but this makes mutter think that a
screen-cast is active, causing gnome-shell to display an icon indicating that
the screen is being shared and causing animations to be disabled. The code
also required unsafe mode to be set in order to work, since it needed the
org.gnome.Shell.Introspect interface, which is normally locked down. We will
need a better API on the mutter/gnome-shell side in order to do this properly.

Also remove the libei dependency, since it is not needed by any other code.

Fixes https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/7438