atk-adaptor: Return error if maximum child count exceeded
    commit 13be6f0edc1d30ef884caf1182dc46f6794bdef8
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Jun 10 14:53:12 2024 +0200

        atk-adaptor: Limit amount of children to return for GetChildren

introduced a maximum amount of children to be returned
by the "GetChildren" method.

Instead of returning that amount of children when
the object contains more, return an error instead.

This makes it clear that an unusual condition occured
that ATs need to handle explicitly, instead of potentially
assuming that the 65536 children that were returned previously
were all children.

See also discussion in merge request !159 [1].

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/merge_requests/159