Accessible.xml: Mention other use cases for AccessibleId
Adust the doc for `AccessibleId` to mention that this can
also be used for reliable identification by ATs, not just
for automated testing.

This is also in line with the existing documentation for
`atk_object_set_accessible_id` (s. atk/atkobject.c) which
mentions:

> Such an ID will be available for
> instance to identify a given well-known accessible object for tailored screen
> reading, or for automatic regression testing.

One example use case could be reliable identification
of LibreOffice's spellcheck dialog by the Orca screen
reader, see [1].

[1] https://bugs.documentfoundation.org/show_bug.cgi?id=155447