Fix possible double unref in add_accessible_from_iter
If the new accessible has the same child index as an existing accessible in
the parent's list of children, then the current behavior is to overwrite the
old object. But, if we manually unref the old object and release the last
reference, then the accessible will be disposed when it is not yet removed
from the parent's list of children, so it will be unrefed a second time in
atspi_accessible_dispose, resulting in a use after free. This might only be
a theoretical issue, since, currently, the path -> accessible hashmap is
likely to also contain a reference to the child, but, regardless, this
change will become necessary with a follow-up commit in the main branch.

Helps #172