Don't keep references to uncached objects in the application's hash table
This can cause a memory leak in case of a long-running application, since
references to accessibles are kept alive in the hash even after the
accessibles are no longer present in the hierarchy, until the application is
destroyed. Instead, remove an object from its application's hash when the
object is disposed. If we have cached data for an object via GetItems or
AddAccessible, then keep a reference to the object to ensure that the cached
data stays alive. Add a private field to track whether we have such a
reference. If we receive a RemoveAccessible, then we dispose the object, and
remove the cache ref there.

Fixes #172