Use the appropriate annotations for callbacks
The (closure) and (destroy) annotations are meant to be used on the
callback argument, and refer to the arguments that contain the data to
be passed to the callback, and the notification function, respectively.

The unary (closure) annotation is meant to be used to describe the user
data on a callback type definition.

The "notified" scope is needed for callbacks that have a destroy
notification function; if there is no notification function, and the
callback is invoked after the function returns, then the correct scope
is "async".

See: https://gi.readthedocs.io/en/latest/annotations/giannotations.html#support-for-gobject-closures