atk-adaptor: Have the SpiCache keep its own weak ref
This makes the code less convoluted. Previously, SpiCache relied on the
code in accessible-register.c to send a signal when the object went away. At
least in theory, this could cause a dangling pointer if the root
accessible changes or goes away, since it has a fixed DBus object path and
is thus not registered in SpiRegister. We have a couple of crashes involving
AT-SPI that are hard to reproduce or debug (#176, #178), and it seems
better to simplify the code.

The SpiRegister now only emits object-deregistered when an object is
marked defunct, not when it is deregistered via the weak ref. SpiCache is
its only consumer and is now notified directly when an AtkObject is being
finalized.