device-a11y-manager: Request events for non-grabbed keys
After some discussions, it was decided that having an explicit
request for monitoring the keyboard is a better design, so use it.

This keeps the previous semantics for AT-SPI2 clients, if better ways of
doing what the screen readers are doing with the key events now is found,
we might easily add a way for them to unsubscribe from the non-grabbed events.