xml: Add version property to AT-SPI interfaces
It is currently impossible to negotiate additions to the AT-SPI
interfaces: if a new symbol is added (property, signal, or method), or a
new member is added to an enumeration, there's no way for toolkits to
communicate that information to the ATs.

Starting from this point, any addition to the existing interfaces will
result in the version property being increased by one.

Unfortunately, the Application interface already has a version property
which maps, for some reason, to the toolkit version; instead of breaking
it, since it would involved changing the type of the property as well,
we:

- deprecate Application:Version
- add Application:ToolkitVersion, to match Application:ToolkitName
- add Application:InterfaceVersion