AtspiDeviceX11: Suppress window events following key grabs when needed
AtspiDeviceX11 attempts to track the input focus and associate key grabs with
the focused window. This prevents the window from losing focus when a grab is
activated with some, but not all window managers (ie, not with I3 or
Ratpoison). When we detect a key press from a grabbed key, we now start a
timer, and, if we see a window event within 20ms of the relevant key press,
then we ignore it. This is not a proper fix, but hopefully it will be an
improvement over the current behavior. We only want to enable this hack
where it is needed, so, as a proxy, check that XDG_CURRENT_DESKTOP is unset.

Closes https://gitlab.gnome.org/GNOME/orca/-/issues/434

Fix up last commit