atk-adaptor: Handle Application iface in spi_object_has_dbus_interface
Add handling for the At-SPI Application interface
to `spi_object_has_dbus_interface` that uses the
same logic to decide whether an object supports that
interface as `spi_object_append_interfaces` already does,
i.e. an AtkObject of role `ATK_ROLE_APPLICATION` supports
that interface.

This fixes an inconsistency with the object reporting that
interface as supported, but then triggering an error when
the interface's properties are queried.

This was reported in the #a11y:gnome.org matrix chat.

Example using Gedit without this commit in place:

    $ busctl call --address='unix:path=/run/user/1000/at-spi/bus_1,guid=1b4dd2a60e220b4112a58fcc682ac20f' ':1.160' '/org/a11y/atspi/accessible/root' 'org.a11y.atspi.Accessible' 'GetInterfaces'
    as 3 "org.a11y.atspi.Accessible" "org.a11y.atspi.Application" "org.a11y.atspi.Collection"
    $ busctl call --address='unix:path=/run/user/1000/at-spi/bus_1,guid=1b4dd2a60e220b4112a58fcc682ac20f' ':1.160' '/org/a11y/atspi/accessible/root' 'org.freedesktop.DBus.Properties' 'GetAll' 's' 'org.a11y.atspi.Application'
    Call failed: Property unavailable

With this commit in place:

    $ busctl call --address='unix:path=/run/user/1000/at-spi/bus_1,guid=1b4dd2a60e220b4112a58fcc682ac20f' ':1.165' '/org/a11y/atspi/accessible/root' 'org.a11y.atspi.Accessible' 'GetInterfaces'
    as 3 "org.a11y.atspi.Accessible" "org.a11y.atspi.Application" "org.a11y.atspi.Collection"
    $ busctl call --address='unix:path=/run/user/1000/at-spi/bus_1,guid=1b4dd2a60e220b4112a58fcc682ac20f' ':1.165' '/org/a11y/atspi/accessible/root' 'org.freedesktop.DBus.Properties' 'GetAll' 's' 'org.a11y.atspi.Application'
    a{sv} 4 "ToolkitName" s "gtk" "AtspiVersion" s "2.1" "Version" s "3.24.49" "Id" i 133