atk-adaptor: Reconnect when the bus launcher reappears
On Wayland sessions, at least on GNOME, the AT-SPI bus is taken down when
the screen reader is disabled, so, if the screen reader is subsequently
re-enabled, then the at-spi-bus-launcher is restarted, but existing
applications (including gnome-shell) are now inaccessible because they don't
know about the new bus. Now libatk-bridge connects to the session bus and
listens for NameOwnerChanged signals for org.a11y.Bus, and, when the name
appears, it reconnects to the accessibility bus. This required refactoring
atk_bridge_adaptor_init and atk_bridge_adaptor_cleanup to separate out the
bus-related code so that it can be used elsewhere.

Helps #193