atk-bridge: Don't crash when requesting a plug if not activated
This isn't really a complete fix. The format of the plug string relies on
an application's name on the accessibility bus, but, if, for instance, the
screen reader is turned off and back on, then it is possible for the
accessibility bus to be taken down and later recreated (see #193). We should
perhaps construct some sort of well-known name that can be set on the bus,
while ensuring that we don't conflict with any other application.

But this is a quick fix that is at least an improvement over a crash.

Closes #198