atk-adaptor: Don't lease objects multiple times
If spi_leasing_take is called on an object that is already leased, then
remove the original entry from the queue, and add a new one with a refreshed
timestamp. Add a hash table to store queue links for objects to allow fast
searching. This shouldn't fix or break anything but avoids having the
leasing code hold multiple refs on an object and might simplify debugging
ref-related issues in the future.

Related to https://gitlab.gnome.org/GNOME/at-spi2-core/-/issues/176