atk: Handle Accessible + Collection iface in match rule
When checking whether an AtkObject implements a given
interface specified by a string, also handle
"Accessible" and "Collection" instead of considering
those as unsupported

Since all AtkObjects implement these interfaces
(see also spi_object_append_interfaces),
return true unless a null AtkObject param is passed.

This fixes querying the a11y tree using the AT-SPI
Collection interface when one of the 2 previously
ignored interface names is specified as a search
criterion.

Example:

Without this commit in place, starting gtk3-demo,
selecting the root application node in Accerciser
and querying for the first descendant that has the
visible state and implements the Action interface
would give the expected result, a list containing
one object that fulfils these criteria:

    In [44]: collection = acc.queryCollection()
    rule = collection.createMatchRule(Atspi.StateSet(Atspi.StateType.VISIBLE), Atspi.CollectionMatchType.ALL,
                                      [], Atspi.CollectionMatchType.ALL,
                                      [], Atspi.CollectionMatchType.ALL,
                                      ["Action"], Atspi.CollectionMatchType.ALL,
                                      False)
    collection.getMatches(rule, Atspi.CollectionSortOrder.CANONICAL, 1, True)
    Out[44]: [<Atspi.Accessible object at 0x7ff92f3ba540 (AtspiAccessible at 0xdc2c620)>]

That object also reports to support the Accessible and
Collection AT-SPI interfaces:

    In [45]: collection.getMatches(rule, Atspi.CollectionSortOrder.CANONICAL, 1, True)[0].get_interfaces()
    Out[45]: ['Accessible', 'Action', 'Collection', 'Component', 'Image']

However, using a rule that would require the Accessible interface
in addition to what the previous rule required would not return
a list containing that object, but an empty list:

    In [46]: collection = acc.queryCollection()
    rule = collection.createMatchRule(Atspi.StateSet(Atspi.StateType.VISIBLE), Atspi.CollectionMatchType.ALL,
                                      [], Atspi.CollectionMatchType.ALL,
                                      [], Atspi.CollectionMatchType.ALL,
                                      ["Action", "Accessible"], Atspi.CollectionMatchType.ALL,
                                      False)
    collection.getMatches(rule, Atspi.CollectionSortOrder.CANONICAL, 1, True)
    Out[46]: []

With this commit in place, the object (or its counterpart
from the next run of gtk3-demo) is matched and returned
as expected:

    In [60]: collection = acc.queryCollection()
    rule = collection.createMatchRule(Atspi.StateSet(Atspi.StateType.VISIBLE), Atspi.CollectionMatchType.ALL,
                                      [], Atspi.CollectionMatchType.ALL,
                                      [], Atspi.CollectionMatchType.ALL,
                                      ["Action", "Accessible"], Atspi.CollectionMatchType.ALL,
                                      False)
    collection.getMatches(rule, Atspi.CollectionSortOrder.CANONICAL, 1, True)
    Out[60]: [<Atspi.Accessible object at 0x7ff92fb109c0 (AtspiAccessible at 0xdba90a0)>]