atk: Drop always TRUE recurse param for Collection adaptor functions
The recurse param is always passed as TRUE
for both, sort_order_canonical and query_exec.

Recursive querying only happens when the
traverse param is TRUE in addition.

Therefore, that param is the one that
actually decides whether or not the tree
is queried recursively.

Drop the recurse param altogether to make
that more obvious and the code more readable.