AtspiDeviceX11: Fix key events with non-default keyboard layouts
Use XkbGetState to try to find the active group, and pass this to
XkbKeycodeToKeysym to get the correct keysym. Do not call XLookupString,
since it might return the wrong keysym and text in set-ups with multiple
keyboard layouts. Consumers now need to derive text from the keysym (orca
already does this if the event's text is empty).

This is essentially the patch from ilusha.paschuk@gmail.com posted to
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1111485, with fixes for
style and warnings.

Closes #185
Closes #201